@@ -40,6 +40,11 @@ public class SupabaseClientFactory {
 
     /**
      * Creates a factory with explicit credentials.
+     * @param host the database host address
+     * @param port the database port number
+     * @param database the database name
+     * @param user the database username
+     * @param password the database password
      */
     public SupabaseClientFactory(final String host, final String port, final String database,
                                  final String user, final String password) {
@@ -57,7 +62,9 @@ public class SupabaseClientFactory {
      * @throws SQLException if connection fails
      */
     public Connection createConnection() throws SQLException {
-        final String jdbcUrl = String.format("jdbc:postgresql://%s:%s/%s?prepareThreshold=0&preparedStatementCacheQueries=0",
+        System.out.println("[SupabaseClientFactory] Opening new database connection...");
+        final String jdbcUrl = String.format("jdbc:postgresql://%s:%s/%s?\n"
+                + "prepareThreshold=0&preparedStatementCacheQueries=0",
             host, port, database);
         return DriverManager.getConnection(jdbcUrl, user, password);
     }
@@ -65,13 +72,23 @@ public class SupabaseClientFactory {
     /**
      * Gets environment variable or returns default value.
      * Checks both system environment and system properties (.env file support).
+     * @param envVar the environment variable set in the .env
+     * @param defaultValue the default variable to use in place of the envVar
+     * @return the .env value if it exists, default value if it doesn't
      */
     private String getEnvOrDefault(final String envVar, final String defaultValue) {
         String value = System.getenv(envVar);
+        final String res;
         if (value == null) {
             value = System.getProperty(envVar);
         }
-        return value != null ? value : defaultValue;
+        if (value != null) {
+            res = value;
+        }
+        else {
+            res = defaultValue;
+        }
+        return res;
     }
 }
 
