@@ -10,6 +10,8 @@ import stakemate.entity.Game;
 /**
  * Interactor for the FetchGames use case.
  * Orchestrates the flow: API -> Adapter -> Repository -> Presenter
+ *
+ * <p>
  */
 public class FetchGamesInteractor implements FetchGamesInputBoundary {
 
@@ -36,51 +38,77 @@ public class FetchGamesInteractor implements FetchGamesInputBoundary {
         presenter.presentFetchInProgress();
 
         try {
-            // Validate sport parameter
-            if (sport == null || sport.trim().isEmpty()) {
-                presenter.presentFetchError("Sport parameter is required");
-                return;
-            }
+            processFetchAndUpdate(sport, region, dateFrom);
+        }
+        catch (final ApiException ex) {
+            presenter.presentFetchError("API error: " + ex.getMessage());
+        }
+        catch (final RepositoryException ex) {
+            presenter.presentFetchError("Database error: " + ex.getMessage());
+        }
+        catch (final Exception ex) {
+            presenter.presentFetchError("Unexpected error: " + ex.getMessage());
+        }
+    }
 
-            // Use today if dateFrom is null
-            final LocalDate effectiveDate = dateFrom != null ? dateFrom : LocalDate.now();
-            final List<OddsApiEvent> events = apiGateway.fetchEvents(sport, region, effectiveDate);
+    /**
+     * Processes the fetch and update logic.
+     *
+     * @param sport The sport to fetch
+     * @param region The region to filter by
+     * @param dateFrom The date to fetch from
+     * @throws ApiException if API call fails
+     * @throws RepositoryException if database operation fails
+     */
+    private void processFetchAndUpdate(final String sport, final String region, final LocalDate dateFrom)
+            throws ApiException, RepositoryException {
+        // Check if API gateway is configured
+        if (apiGateway == null) {
+            presenter.presentFetchError("API gateway not configured. Please set ODDS_API_KEY.");
+            return;
+        }
 
-            if (events.isEmpty()) {
-                presenter.presentFetchSuccess(new FetchGamesResponseModel(
-                    0, 0, sport, "No events found for the specified criteria."
-                ));
-                return;
-            }
+        // Validate sport parameter
+        if (sport == null || sport.trim().isEmpty()) {
+            presenter.presentFetchError("Sport parameter is required");
+            return;
+        }
 
-            // Step 2: Convert API events to Game entities
-            final List<Game> games = responseAdapter.convertToGames(events);
+        // Use today if dateFrom is null
+        final LocalDate effectiveDate;
+        if (dateFrom != null) {
+            effectiveDate = dateFrom;
+        }
+        else {
+            effectiveDate = LocalDate.now();
+        }
 
-            if (games.isEmpty()) {
-                presenter.presentFetchSuccess(new FetchGamesResponseModel(
-                    events.size(), 0, sport, "Fetched events but none could be converted to games."
-                ));
-                return;
-            }
+        final List<OddsApiEvent> events = apiGateway.fetchEvents(sport, region, effectiveDate);
 
-            // Step 3: Normalize and validate game data
-            final List<Game> validGames = normalizeAndValidateGames(games);
-            gameRepository.upsertGames(validGames);
-            final String message = String.format("Successfully fetched and saved %d games.", validGames.size());
+        if (events.isEmpty()) {
             presenter.presentFetchSuccess(new FetchGamesResponseModel(
-                events.size(), validGames.size(), sport, message
+                0, 0, sport, "No events found for the specified criteria."
             ));
+            return;
+        }
 
+        // Step 2: Convert API events to Game entities
+        final List<Game> games = responseAdapter.convertToGames(events);
+
+        if (games.isEmpty()) {
+            presenter.presentFetchSuccess(new FetchGamesResponseModel(
+                events.size(), 0, sport, "Fetched events but none could be converted to games."
+            ));
+            return;
         }
-        catch (final ApiException e) {
-            presenter.presentFetchError("API error: " + e.getMessage());
-        }
-        catch (final RepositoryException e) {
-            presenter.presentFetchError("Database error: " + e.getMessage());
-        }
-        catch (final Exception e) {
-            presenter.presentFetchError("Unexpected error: " + e.getMessage());
-        }
+
+        // Step 3: Normalize and validate game data
+        final List<Game> validGames = normalizeAndValidateGames(games);
+        gameRepository.upsertGames(validGames);
+        final String message = String.format("Successfully fetched and saved %d games.", validGames.size());
+        presenter.presentFetchSuccess(new FetchGamesResponseModel(
+            events.size(), validGames.size(), sport, message
+        ));
     }
 
     @Override
@@ -94,40 +122,56 @@ public class FetchGamesInteractor implements FetchGamesInputBoundary {
             final List<Game> games = gameRepository.searchGames(query);
             presenter.presentSearchResults(games, query);
         }
-        catch (final RepositoryException e) {
-            presenter.presentFetchError("Search failed: " + e.getMessage());
+        catch (final RepositoryException ex) {
+            presenter.presentFetchError("Search failed: " + ex.getMessage());
         }
     }
 
     /**
      * Normalizes and validates game data before saving.
      * Filters out invalid games and ensures data consistency.
+     *
+     * @param games The list of games to validate
+     * @return List of valid games
      */
     private List<Game> normalizeAndValidateGames(final List<Game> games) {
         final List<Game> validGames = new ArrayList<>();
 
         for (final Game game : games) {
-            // Validate required fields
-            if (game.getId() == null ||
-                game.getMarketId() == null ||
-                game.getGameTime() == null ||
-                game.getTeamA() == null || game.getTeamA().trim().isEmpty() ||
-                game.getTeamB() == null || game.getTeamB().trim().isEmpty() ||
-                game.getSport() == null || game.getSport().trim().isEmpty()) {
-                // Skip invalid games
-                continue;
+            if (isValidGame(game) && isFutureGame(game)) {
+                validGames.add(game);
             }
-
-            // Only include future games (game_time >= today)
-            if (game.getGameTime().isBefore(java.time.LocalDateTime.now())) {
-                // Skip past games
-                continue;
-            }
-
-            validGames.add(game);
         }
 
         return validGames;
+    }
+
+    /**
+     * Validates if a game has all required fields.
+     *
+     * @param game The game to validate
+     * @return true if game is valid, false otherwise
+     */
+    private boolean isValidGame(final Game game) {
+        return game.getId() != null
+                && game.getMarketId() != null
+                && game.getGameTime() != null
+                && game.getTeamA() != null
+                && !game.getTeamA().trim().isEmpty()
+                && game.getTeamB() != null
+                && !game.getTeamB().trim().isEmpty()
+                && game.getSport() != null
+                && !game.getSport().trim().isEmpty();
+    }
+
+    /**
+     * Checks if a game is in the future.
+     *
+     * @param game The game to check
+     * @return true if game time is in the future, false otherwise
+     */
+    private boolean isFutureGame(final Game game) {
+        return !game.getGameTime().isBefore(java.time.LocalDateTime.now());
     }
 }
 
