@@ -10,18 +10,22 @@ public class Bet {
     private final String username;
     private final String marketId;
     private final Side side;
-    private final double stake;
-    private final double price;
-    private final Boolean won;
-    private final Boolean settled;
+    private final double stake; // corresponds to positions.amount
+    private final double price; // price/odds
+    private final Boolean won; // true = won, false = lost, null = not settled yet
+    private final Boolean settled; // true = already settled, false = not settled
+    private final String teamName; // Resolved team name (e.g. "Lakers")
+    private final java.time.Instant updatedAt; // Timestamp of last update
 
     public Bet(String username,
-               String marketId,
-               Side side,
-               double stake,
-               double price,
-               Boolean won,
-               Boolean settled) {
+            String marketId,
+            Side side,
+            double stake,
+            double price,
+            Boolean won,
+            Boolean settled,
+            String teamName,
+            java.time.Instant updatedAt) {
 
         this.username = username;
         this.marketId = marketId;
@@ -30,6 +34,29 @@ public class Bet {
         this.price = price;
         this.won = won;
         this.settled = settled;
+        this.teamName = teamName;
+        this.updatedAt = updatedAt;
+    }
+
+    public Bet(String username,
+            String marketId,
+            Side side,
+            double stake,
+            double price,
+            Boolean won,
+            Boolean settled,
+            String teamName) {
+        this(username, marketId, side, stake, price, won, settled, teamName, java.time.Instant.now());
+    }
+
+    public Bet(String username,
+            String marketId,
+            Side side,
+            double stake,
+            double price,
+            Boolean won,
+            Boolean settled) {
+        this(username, marketId, side, stake, price, won, settled, null, java.time.Instant.now());
     }
 
     public String getUsername() {
@@ -58,5 +85,13 @@ public class Bet {
 
     public Boolean isSettled() {
         return settled;
+    }
+
+    public String getTeamName() {
+        return teamName;
+    }
+
+    public java.time.Instant getUpdatedAt() {
+        return updatedAt;
     }
 }
