@@ -4,13 +4,24 @@ import java.util.ArrayList;
 import java.util.List;
 
 import stakemate.entity.User;
+import stakemate.use_case.settle_market.Bet;
+import stakemate.use_case.view_profile.strategy.BetComparator;
 
+/**
+ * Interactor for the View Profile Use Case.
+ */
 public class ViewProfileInteractor implements ViewProfileInputBoundary {
     private final ViewProfileUserDataAccessInterface userDataAccess;
     private final ViewProfileOutputBoundary outputBoundary;
 
+    /**
+     * Constructs a ViewProfileInteractor.
+     * 
+     * @param userDataAccess the data access interface.
+     * @param outputBoundary the output boundary.
+     */
     public ViewProfileInteractor(final ViewProfileUserDataAccessInterface userDataAccess,
-                                 final ViewProfileOutputBoundary outputBoundary) {
+            final ViewProfileOutputBoundary outputBoundary) {
         this.userDataAccess = userDataAccess;
         this.outputBoundary = outputBoundary;
     }
@@ -25,24 +36,37 @@ public class ViewProfileInteractor implements ViewProfileInputBoundary {
             return;
         }
 
-        // Hardcoded exemplar data as requested
-        final List<String[]> openPositions = new ArrayList<>();
-        // Won}
-        openPositions.add(new String[]{
-            "Lakers vs Warriors", "Lakers", "0.60", "100", "60.00", "40.00"
-        });
+        final List<Bet> bets = userDataAccess.getPositionsByUsername(username);
 
-        final List<String[]> historicalPositions = new ArrayList<>();
-        historicalPositions.add(new String[]{
-            "Knicks vs Celtics", "Celtics", "0.40", "50", "30.00"
-        });
+        final List<Bet> openBets = new ArrayList<>();
+        final List<Bet> historicalBets = new ArrayList<>();
+
+        for (final Bet bet : bets) {
+            if (Boolean.TRUE.equals(bet.isSettled())) {
+                historicalBets.add(bet);
+            } else {
+                openBets.add(bet);
+            }
+        }
+
+        // Sort Open Bets using Strategy
+        final BetComparator openStrategy = inputData.getOpenSortStrategy();
+        if (openStrategy != null) {
+            openBets.sort(openStrategy);
+        }
+
+        // Sort Historical Bets using Strategy
+        final BetComparator histStrategy = inputData.getHistoricalSortStrategy();
+        if (histStrategy != null) {
+            historicalBets.sort(histStrategy);
+        }
 
         final ViewProfileOutputData outputData = new ViewProfileOutputData(
-            user.getUsername(),
-            user.getBalance(),
-            0, // PnL set to 0 for now
-            openPositions,
-            historicalPositions);
+                user.getUsername(),
+                user.getBalance(),
+                user.getBalance() - 10000, // PnL = current balance - starting balance (10,000)
+                openBets,
+                historicalBets);
 
         outputBoundary.presentProfile(outputData);
     }
