@@ -1,17 +1,32 @@
 package stakemate.use_case.view_profile;
 
 import java.util.List;
+import stakemate.use_case.settle_market.Bet;
 
+/**
+ * Output Data for the View Profile Use Case.
+ */
 public class ViewProfileOutputData {
     private final String username;
-    private final int balance;
-    private final int pnl;
-    private final List<String[]> openPositions;
-    private final List<String[]> historicalPositions;
+    private final double balance;
+    private final double pnl;
+    private final List<Bet> openPositions;
+    private final List<Bet> historicalPositions;
 
-    public ViewProfileOutputData(final String username, final int balance, final int pnl,
-                                 final List<String[]> openPositions,
-                                 final List<String[]> historicalPositions) {
+    /**
+     * Constructs a ViewProfileOutputData.
+     * 
+     * @param username            the username.
+     * @param balance             the balance.
+     * @param pnl                 the pnl.
+     * @param openPositions       the open positions.
+     * @param historicalPositions the historical positions.
+     */
+    public ViewProfileOutputData(final String username,
+            final double balance,
+            final double pnl,
+            final List<Bet> openPositions,
+            final List<Bet> historicalPositions) {
         this.username = username;
         this.balance = balance;
         this.pnl = pnl;
@@ -19,23 +34,48 @@ public class ViewProfileOutputData {
         this.historicalPositions = historicalPositions;
     }
 
+    /**
+     * Gets the username.
+     * 
+     * @return the username.
+     */
     public String getUsername() {
         return username;
     }
 
-    public int getBalance() {
+    /**
+     * Gets the balance.
+     * 
+     * @return the balance.
+     */
+    public double getBalance() {
         return balance;
     }
 
-    public int getPnl() {
+    /**
+     * Gets the PnL.
+     * 
+     * @return the PnL.
+     */
+    public double getPnl() {
         return pnl;
     }
 
-    public List<String[]> getOpenPositions() {
+    /**
+     * Gets the open positions.
+     * 
+     * @return the open positions.
+     */
+    public List<Bet> getOpenPositions() {
         return openPositions;
     }
 
-    public List<String[]> getHistoricalPositions() {
+    /**
+     * Gets the historical positions.
+     * 
+     * @return the historical positions.
+     */
+    public List<Bet> getHistoricalPositions() {
         return historicalPositions;
     }
 }
