@@ -3,8 +3,6 @@ package stakemate.view;
 import java.awt.BorderLayout;
 import java.awt.Color;
 import java.awt.FlowLayout;
-import java.util.ArrayList;
-import java.util.Comparator;
 import java.util.List;
 
 import javax.swing.BorderFactory;
@@ -21,15 +19,16 @@ import javax.swing.JTable;
 import javax.swing.ListSelectionModel;
 import javax.swing.SwingConstants;
 import javax.swing.Timer;
-import javax.swing.table.AbstractTableModel;
+import javax.swing.event.ListSelectionEvent;
+import javax.swing.event.ListSelectionListener;
 import javax.swing.table.DefaultTableCellRenderer;
 import javax.swing.table.TableCellRenderer;
 
 import stakemate.app.StakeMateApp;
-import stakemate.entity.OrderBook;
-import stakemate.entity.OrderBookEntry;
 import stakemate.interface_adapter.controllers.SettleMarketController;
 import stakemate.interface_adapter.view_comments.PostCommentController;
+import stakemate.interface_adapter.view_comments.ViewCommentsController;
+import stakemate.interface_adapter.view_live.LiveMatchesController;
 import stakemate.interface_adapter.view_market.MarketsView;
 import stakemate.interface_adapter.view_market.SettleMarketView;
 import stakemate.interface_adapter.view_market.ViewMarketController;
@@ -38,28 +37,21 @@ import stakemate.use_case.view_market.MarketsResponseModel;
 import stakemate.use_case.view_market.MatchSummary;
 import stakemate.use_case.view_market.OrderBookResponseModel;
 
-/**
- * The main frame for viewing markets and order books.
- */
-// -@cs[ClassDataAbstractionCoupling] Swing UI classes require many dependencies to function.
-// -@cs[ClassFanOutComplexity] View layer inevitably depends on many Swing and App components.
+// -@cs[ClassDataAbstractionCoupling] UI class requires many dependencies.
+// -@cs[ClassFanOutComplexity] UI class requires many dependencies.
 public class MarketsFrame extends JFrame implements MarketsView, SettleMarketView {
 
     private static final String EMPTY_TEXT = " ";
     private static final String ERROR_TITLE = "Error";
+    private static final String SELECT_MARKET_MSG = "Select a market to see orders.";
     private static final int WINDOW_WIDTH = 1000;
     private static final int WINDOW_HEIGHT = 600;
     private static final int BORDER_PADDING = 8;
-    private static final int GAP = 4;
     private static final int DIVIDER_LOC = 280;
     private static final int REFRESH_INTERVAL = 5000;
     private static final int DARK_GREEN_G = 100;
-
-    // Table Column Constants
-    private static final int COL_BID_QTY = 0;
-    private static final int COL_BID_PRICE = 1;
-    private static final int COL_ASK_PRICE = 2;
-    private static final int COL_ASK_QTY = 3;
+    private static final int ROW_HEIGHT = 24;
+    private static final int GAP = 4;
 
     private final DefaultListModel<MatchSummary> matchesListModel = new DefaultListModel<>();
     private final JList<MatchSummary> matchesList = new JList<>(matchesListModel);
@@ -71,25 +63,39 @@ public class MarketsFrame extends JFrame implements MarketsView, SettleMarketVie
 
     private final OrderBookTableModel orderBookTableModel = new OrderBookTableModel();
     private final JTable orderBookTable = new JTable(orderBookTableModel);
-    private final JLabel orderBookEmptyLabel = new JLabel("Select a market to see orders.");
+    private final JLabel orderBookEmptyLabel = new JLabel(SELECT_MARKET_MSG);
 
-    private final JLabel statusLabel = new JLabel(EMPTY_TEXT);
+    private final JLabel statusLabel = new JLabel(" ");
+
+    // Buttons
     private final JButton buyButton = new JButton("Buy");
     private final JButton sellButton = new JButton("Sell");
+
+    private final JButton refreshButton = new JButton("Refresh");
     private final JButton myProfileButton = new JButton("My Profile");
+    private final JButton liveMatchesButton = new JButton("Live Matches");
     private final JButton settleButton = new JButton("Settle (Demo)");
+
+    // Place Order Button
+    private final JButton placeOrderButton = new JButton("Place Order");
     private final CommentsPanel commentsPanel = new CommentsPanel();
 
-    private ViewMarketController viewController;
+    private ViewMarketController controller;
     private SettleMarketController settleMarketController;
-    private stakemate.interface_adapter.view_comments.ViewCommentsController viewCommentsController;
-    private stakemate.interface_adapter.view_comments.PostCommentController postCommentController;
+    private ViewCommentsController viewCommentsController;
+    private PostCommentController postCommentController;
     private MarketSummary currentlySelectedMarket;
     private ProfileFrame profileFrame;
     private stakemate.interface_adapter.view_profile.ViewProfileController profileController;
     private String currentUser;
+    private LiveMatchesFrame liveMatchesFrame;
+    private LiveMatchesController liveMatchesController;
 
     private Timer autoRefreshTimer;
+
+    // Listeners stored as fields to allow detachment during updates
+    private ListSelectionListener matchesSelectionListener;
+    private ListSelectionListener marketsSelectionListener;
 
     public MarketsFrame() {
         super("StakeMate - Markets & Order Book");
@@ -97,46 +103,35 @@ public class MarketsFrame extends JFrame implements MarketsView, SettleMarketVie
     }
 
     /**
-     * Sets the ViewMarketController.
+     * Sets the ViewMarketController and initializes hooks.
      *
-     * @param controller The controller for market operations.
+     * @param controllerArg The controller for market operations.
      */
-    public void setController(final ViewMarketController controller) {
-        this.viewController = controller;
+    public void setController(final ViewMarketController controllerArg) {
+        this.controller = controllerArg;
         hookEvents();
-        startAutoRefresh();
+        if (autoRefreshTimer != null && !autoRefreshTimer.isRunning()) {
+            autoRefreshTimer.start();
+        }
+        if (controller != null) {
+            controller.refresh();
+        }
+    }
+
+    public void setSettleMarketController(final SettleMarketController controllerArg) {
+        this.settleMarketController = controllerArg;
+    }
+
+    public void setViewCommentsController(final ViewCommentsController controllerArg) {
+        this.viewCommentsController = controllerArg;
+    }
+
+    public void setPostCommentController(final PostCommentController controllerArg) {
+        this.postCommentController = controllerArg;
     }
 
     /**
-     * Sets the SettleMarketController.
-     *
-     * @param controller The controller for settlement operations.
-     */
-    public void setSettleMarketController(final SettleMarketController controller) {
-        this.settleMarketController = controller;
-    }
-
-    /**
-     * Sets the ViewComment Controller.
-     *
-     * @param controller The controller for comments
-     */
-    public void setViewCommentsController(stakemate.interface_adapter.view_comments.ViewCommentsController controller) {
-        this.viewCommentsController = controller;
-    }
-
-    /**
-     * Sets the PostCommentController.
-     *
-     * @param controller The controller for posting comments
-     */
-    public void setPostCommentController(PostCommentController controller) {
-        this.postCommentController = controller;
-    }
-
-    /**
-     * Sets up the comments panel.
-     *
+     * Wires up the comments panel with controllers.
      */
     public void wireCommentsPanel() {
         if (postCommentController != null && viewCommentsController != null) {
@@ -145,30 +140,23 @@ public class MarketsFrame extends JFrame implements MarketsView, SettleMarketVie
         }
     }
 
-    /**
-     * Sets the ProfileFrame for navigation.
-     *
-     * @param profileFrame The user profile window.
-     */
     public void setProfileFrame(final ProfileFrame profileFrame) {
         this.profileFrame = profileFrame;
     }
 
-    /**
-     * Sets the ProfileController.
-     *
-     * @param profileController The controller for profile operations.
-     */
     public void setProfileController(
         final stakemate.interface_adapter.view_profile.ViewProfileController profileController) {
         this.profileController = profileController;
     }
 
-    /**
-     * Sets the currently logged-in user.
-     *
-     * @param username The username.
-     */
+    public void setLiveMatchesFrame(final LiveMatchesFrame liveMatchesFrame) {
+        this.liveMatchesFrame = liveMatchesFrame;
+    }
+
+    public void setLiveMatchesController(final LiveMatchesController liveMatchesController) {
+        this.liveMatchesController = liveMatchesController;
+    }
+
     public void setLoggedInUser(final String username) {
         this.currentUser = username;
     }
@@ -182,7 +170,10 @@ public class MarketsFrame extends JFrame implements MarketsView, SettleMarketVie
         root.setBorder(BorderFactory.createEmptyBorder(
             BORDER_PADDING, BORDER_PADDING, BORDER_PADDING, BORDER_PADDING));
 
-        final JPanel topBar = createTopPanel();
+        // Create Top Panel (inlined)
+        final JPanel topBar = new JPanel(new FlowLayout(FlowLayout.RIGHT));
+        topBar.add(myProfileButton);
+
         final JSplitPane splitPane = createMainSplitPane();
 
         statusLabel.setHorizontalAlignment(SwingConstants.LEFT);
@@ -196,38 +187,30 @@ public class MarketsFrame extends JFrame implements MarketsView, SettleMarketVie
         setContentPane(root);
     }
 
-    private JPanel createTopPanel() {
-        final JPanel topBar = new JPanel(new FlowLayout(FlowLayout.RIGHT));
-        topBar.add(myProfileButton);
-        return topBar;
-    }
-
     private JSplitPane createMainSplitPane() {
         matchesEmptyLabel.setHorizontalAlignment(SwingConstants.CENTER);
         marketsEmptyLabel.setHorizontalAlignment(SwingConstants.CENTER);
         orderBookEmptyLabel.setHorizontalAlignment(SwingConstants.CENTER);
 
-        // --- Left Panel (Matches) ---
+        // Create Left Panel (inlined)
         final JPanel leftPanel = new JPanel(new BorderLayout(GAP, GAP));
-        final JLabel matchesLabel = new JLabel("Active Matches");
+        final JLabel matchesLabel = new JLabel("Matches");
         leftPanel.add(matchesLabel, BorderLayout.NORTH);
         matchesList.setSelectionMode(ListSelectionModel.SINGLE_SELECTION);
         leftPanel.add(new JScrollPane(matchesList), BorderLayout.CENTER);
-        leftPanel.add(matchesEmptyLabel, BorderLayout.SOUTH);
+        final JPanel matchesSouth = new JPanel(new BorderLayout());
+        matchesSouth.add(matchesEmptyLabel, BorderLayout.NORTH);
+        matchesSouth.add(liveMatchesButton, BorderLayout.SOUTH);
+        leftPanel.add(matchesSouth, BorderLayout.SOUTH);
 
-        // --- Right Panel (Markets + OrderBook + Controls) ---
         final JPanel rightPanel = createRightPanel();
 
-        final JSplitPane splitPane = new JSplitPane(
-            JSplitPane.HORIZONTAL_SPLIT,
-            leftPanel,
-            rightPanel);
+        final JSplitPane splitPane = new JSplitPane(JSplitPane.HORIZONTAL_SPLIT, leftPanel, rightPanel);
         splitPane.setDividerLocation(DIVIDER_LOC);
         return splitPane;
     }
 
     private JPanel createRightPanel() {
-        // --- Middle Top (Markets) ---
         final JPanel marketsPanel = new JPanel(new BorderLayout(GAP, GAP));
         final JLabel marketsLabel = new JLabel("Markets");
         marketsPanel.add(marketsLabel, BorderLayout.NORTH);
@@ -235,82 +218,116 @@ public class MarketsFrame extends JFrame implements MarketsView, SettleMarketVie
         marketsPanel.add(new JScrollPane(marketsList), BorderLayout.CENTER);
         marketsPanel.add(marketsEmptyLabel, BorderLayout.SOUTH);
 
-        // --- Middle Bottom (Order Book) ---
-        final JPanel orderBookPanel = new JPanel(new BorderLayout(GAP, GAP));
-        final JLabel orderBookLabel = new JLabel("Live Order Book");
-        orderBookPanel.add(orderBookLabel, BorderLayout.NORTH);
-        orderBookTable.setFillsViewportHeight(true);
+        final JPanel buySellPanel = new JPanel(new FlowLayout(FlowLayout.RIGHT));
+        buyButton.setEnabled(false);
+        sellButton.setEnabled(false);
+        settleButton.setEnabled(false);
+        buySellPanel.add(placeOrderButton);
+        buySellPanel.add(settleButton);
 
-        // Center align the DATA
+        final JPanel rightPanel = new JPanel(new BorderLayout(GAP, GAP));
+        rightPanel.add(marketsPanel, BorderLayout.NORTH);
+        rightPanel.add(createOrderBookPanel(), BorderLayout.CENTER);
+        rightPanel.add(buySellPanel, BorderLayout.SOUTH);
+        return rightPanel;
+    }
+
+    private JPanel createOrderBookPanel() {
+        final JPanel orderBookPanel = new JPanel(new BorderLayout(GAP, GAP));
+        final JLabel orderBookLabel = new JLabel("Order Book");
+        orderBookLabel.setHorizontalAlignment(SwingConstants.CENTER);
+        orderBookPanel.add(orderBookLabel, BorderLayout.NORTH);
+
+        orderBookTable.setFillsViewportHeight(true);
+        orderBookTable.setRowHeight(ROW_HEIGHT);
+
         final DefaultTableCellRenderer centerRenderer = new DefaultTableCellRenderer();
         centerRenderer.setHorizontalAlignment(JLabel.CENTER);
         for (int i = 0; i < orderBookTable.getColumnCount(); i++) {
             orderBookTable.getColumnModel().getColumn(i).setCellRenderer(centerRenderer);
         }
 
-        // Center align the HEADERS
-        final TableCellRenderer headerRenderer =
-            orderBookTable.getTableHeader().getDefaultRenderer();
+        final TableCellRenderer headerRenderer = orderBookTable.getTableHeader().getDefaultRenderer();
         if (headerRenderer instanceof DefaultTableCellRenderer) {
-            ((DefaultTableCellRenderer) headerRenderer)
-                .setHorizontalAlignment(JLabel.CENTER);
+            ((DefaultTableCellRenderer) headerRenderer).setHorizontalAlignment(JLabel.CENTER);
         }
 
         orderBookPanel.add(new JScrollPane(orderBookTable), BorderLayout.CENTER);
         orderBookPanel.add(orderBookEmptyLabel, BorderLayout.SOUTH);
-
-        // --- Controls ---
-        final JPanel buySellPanel = new JPanel(new FlowLayout(FlowLayout.RIGHT));
-        buyButton.setEnabled(false);
-        sellButton.setEnabled(false);
-        settleButton.setEnabled(false);
-
-        buySellPanel.add(buyButton);
-        buySellPanel.add(sellButton);
-        buySellPanel.add(settleButton);
-
-        final JPanel rightPanel = new JPanel(new BorderLayout(GAP, GAP));
-        rightPanel.add(marketsPanel, BorderLayout.NORTH);
-        rightPanel.add(orderBookPanel, BorderLayout.CENTER);
-        rightPanel.add(buySellPanel, BorderLayout.SOUTH);
-
-        return rightPanel;
+        return orderBookPanel;
     }
 
     private void hookEvents() {
         autoRefreshTimer = new Timer(REFRESH_INTERVAL, evt -> {
-            if (viewController != null) {
-                viewController.refresh();
+            if (controller != null) {
+                controller.refresh();
+            }
+        });
+
+        refreshButton.addActionListener(evt -> {
+            if (controller != null) {
+                controller.refreshWithApi();
             }
         });
 
         myProfileButton.addActionListener(evt -> openProfile());
 
-        matchesList.addListSelectionListener(evt -> {
-            if (!evt.getValueIsAdjusting() && viewController != null) {
-                final MatchSummary selected = matchesList.getSelectedValue();
-                currentlySelectedMarket = null;
-                viewController.onMatchSelected(selected);
+        liveMatchesButton.addActionListener(evt -> {
+            if (liveMatchesFrame != null && liveMatchesController != null) {
+                liveMatchesFrame.setVisible(true);
+                liveMatchesController.startTracking();
+            }
+            else {
+                JOptionPane.showMessageDialog(this, "Live Matches view not connected.");
             }
         });
 
-        marketsList.addListSelectionListener(evt -> {
-            if (!evt.getValueIsAdjusting()) {
-                final MarketSummary selected = marketsList.getSelectedValue();
-                currentlySelectedMarket = selected;
+        matchesSelectionListener = this::handleMatchSelection;
+        matchesList.addListSelectionListener(matchesSelectionListener);
 
-                if (viewController != null && selected != null) {
-                    viewController.onMarketSelected(selected);
-                }
+        marketsSelectionListener = this::handleMarketSelection;
+        marketsList.addListSelectionListener(marketsSelectionListener);
 
-                // >>> ADD THIS: automatically load comments
-                if (viewCommentsController != null && selected != null) {
-                    viewCommentsController.fetchComments(selected.getId());
-                }
+        placeOrderButton.addActionListener(evt -> openOrderBookPopup());
+
+        settleButton.addActionListener(evt -> {
+            if (settleMarketController != null && currentlySelectedMarket != null) {
+                performSettlement();
             }
         });
+    }
 
-        hookButtons();
+    private void handleMatchSelection(final ListSelectionEvent evt) {
+        if (!evt.getValueIsAdjusting() && controller != null) {
+            final MatchSummary selected = matchesList.getSelectedValue();
+            currentlySelectedMarket = null;
+
+            orderBookTableModel.clear();
+            orderBookEmptyLabel.setText(SELECT_MARKET_MSG);
+            updateButtonStates(false);
+
+            controller.onMatchSelected(selected);
+        }
+    }
+
+    private void handleMarketSelection(final ListSelectionEvent evt) {
+        if (!evt.getValueIsAdjusting()) {
+            final MarketSummary selected = marketsList.getSelectedValue();
+            currentlySelectedMarket = selected;
+
+            if (selected != null) {
+                orderBookTableModel.clear();
+                orderBookEmptyLabel.setText("Loading...");
+            }
+
+            if (controller != null && selected != null) {
+                controller.onMarketSelected(selected);
+            }
+
+            if (viewCommentsController != null && selected != null) {
+                viewCommentsController.fetchComments(selected.getId());
+            }
+        }
     }
 
     private void openProfile() {
@@ -325,42 +342,18 @@ public class MarketsFrame extends JFrame implements MarketsView, SettleMarketVie
         }
     }
 
-    private void hookButtons() {
-        buyButton.addActionListener(evt -> {
-            JOptionPane.showMessageDialog(
-                MarketsFrame.this,
-                "Buy clicked. (Handled by PlaceOrderUseCase)");
-        });
-
-        sellButton.addActionListener(evt -> {
-            JOptionPane.showMessageDialog(
-                MarketsFrame.this,
-                "Sell clicked. (Handled by PlaceOrderUseCase)");
-        });
-
-        settleButton.addActionListener(evt -> performSettlement());
-    }
-
     private void performSettlement() {
-        if (settleMarketController != null && currentlySelectedMarket != null) {
-            final int choice = JOptionPane.showConfirmDialog(
-                MarketsFrame.this,
-                "Demo Tool: Did the HOME team win this market?",
-                "Simulate Settlement",
-                JOptionPane.YES_NO_OPTION);
+        final int choice = JOptionPane.showConfirmDialog(
+            MarketsFrame.this,
+            "Demo Tool: Did the HOME team win this market?",
+            "Simulate Settlement",
+            JOptionPane.YES_NO_OPTION);
 
-            if (choice == JOptionPane.YES_OPTION || choice == JOptionPane.NO_OPTION) {
-                settleMarketController.settleMarket(currentlySelectedMarket.getId());
-            }
+        if (choice == JOptionPane.YES_OPTION) {
+            settleMarketController.settleMarket(currentlySelectedMarket.getId(), true);
         }
-    }
-
-    private void startAutoRefresh() {
-        if (autoRefreshTimer != null && !autoRefreshTimer.isRunning()) {
-            autoRefreshTimer.start();
-        }
-        if (viewController != null) {
-            viewController.refresh();
+        else if (choice == JOptionPane.NO_OPTION) {
+            settleMarketController.settleMarket(currentlySelectedMarket.getId(), false);
         }
     }
 
@@ -368,7 +361,13 @@ public class MarketsFrame extends JFrame implements MarketsView, SettleMarketVie
 
     @Override
     public void showMatches(final List<MatchSummary> matches, final String emptyStateMessage) {
-        final int selectedIndex = matchesList.getSelectedIndex();
+        String selectedId = null;
+        final MatchSummary current = matchesList.getSelectedValue();
+        if (current != null) {
+            selectedId = current.getId();
+        }
+
+        matchesList.removeListSelectionListener(matchesSelectionListener);
 
         matchesListModel.clear();
         if (matches == null || matches.isEmpty()) {
@@ -386,20 +385,26 @@ public class MarketsFrame extends JFrame implements MarketsView, SettleMarketVie
             matchesEmptyLabel.setText(EMPTY_TEXT);
         }
 
-        if (selectedIndex >= 0 && selectedIndex < matchesListModel.size()) {
-            matchesList.setSelectedIndex(selectedIndex);
-        }
+        restoreListSelection(matchesList, selectedId);
+        matchesList.addListSelectionListener(matchesSelectionListener);
     }
 
     @Override
     public void showMarketsForMatch(final MarketsResponseModel responseModel) {
         setTitle("StakeMate - " + responseModel.getMatchTitle());
 
-        final int selectedIndex = marketsList.getSelectedIndex();
+        String selectedId = null;
+        final MarketSummary current = marketsList.getSelectedValue();
+        if (current != null) {
+            selectedId = current.getId();
+        }
+
+        marketsList.removeListSelectionListener(marketsSelectionListener);
 
         marketsListModel.clear();
-        currentlySelectedMarket = null;
+
         if (responseModel.getMarkets() == null || responseModel.getMarkets().isEmpty()) {
+            currentlySelectedMarket = null;
             if (responseModel.getEmptyStateMessage() != null) {
                 marketsEmptyLabel.setText(responseModel.getEmptyStateMessage());
             }
@@ -414,14 +419,46 @@ public class MarketsFrame extends JFrame implements MarketsView, SettleMarketVie
             marketsEmptyLabel.setText(EMPTY_TEXT);
         }
 
-        if (selectedIndex >= 0 && selectedIndex < marketsListModel.size()) {
-            marketsList.setSelectedIndex(selectedIndex);
+        restoreListSelection(marketsList, selectedId);
+
+        // Update current selection based on list state
+        if (marketsList.getSelectedIndex() != -1) {
             currentlySelectedMarket = marketsList.getSelectedValue();
         }
         else {
+            currentlySelectedMarket = null;
+        }
+
+        if (currentlySelectedMarket == null) {
             orderBookTableModel.clear();
-            orderBookEmptyLabel.setText("Select a market to see orders.");
+            orderBookEmptyLabel.setText(SELECT_MARKET_MSG);
             updateButtonStates(false);
+        }
+        else {
+            updateButtonStates(currentlySelectedMarket.isBuySellEnabled());
+        }
+
+        marketsList.addListSelectionListener(marketsSelectionListener);
+    }
+
+    private <T> void restoreListSelection(final JList<T> list, final String selectedId) {
+        if (selectedId != null) {
+            final javax.swing.ListModel<T> model = list.getModel();
+            for (int i = 0; i < model.getSize(); i++) {
+                final T item = model.getElementAt(i);
+                String itemId = null;
+                if (item instanceof MatchSummary) {
+                    itemId = ((MatchSummary) item).getId();
+                }
+                else if (item instanceof MarketSummary) {
+                    itemId = ((MarketSummary) item).getId();
+                }
+
+                if (selectedId.equals(itemId)) {
+                    list.setSelectedIndex(i);
+                    break;
+                }
+            }
         }
     }
 
@@ -431,6 +468,28 @@ public class MarketsFrame extends JFrame implements MarketsView, SettleMarketVie
             orderBookTableModel.setOrderBook(responseModel.getOrderBook());
         }
 
+        updateStatusLabels(responseModel);
+
+        if (responseModel.isEmpty()) {
+            if (responseModel.getMessage() != null) {
+                orderBookEmptyLabel.setText(responseModel.getMessage());
+            }
+            else {
+                orderBookEmptyLabel.setText("No orders yet");
+            }
+        }
+        else if (!responseModel.isReconnecting()) {
+            orderBookEmptyLabel.setText(EMPTY_TEXT);
+        }
+
+        final boolean enableBuySell = currentlySelectedMarket != null
+            && currentlySelectedMarket.isBuySellEnabled()
+            && !responseModel.isReconnecting();
+
+        updateButtonStates(enableBuySell);
+    }
+
+    private void updateStatusLabels(final OrderBookResponseModel responseModel) {
         if (responseModel.isReconnecting()) {
             if (responseModel.getMessage() != null) {
                 statusLabel.setText(responseModel.getMessage());
@@ -448,39 +507,13 @@ public class MarketsFrame extends JFrame implements MarketsView, SettleMarketVie
             statusLabel.setText("System: Live");
             statusLabel.setForeground(new Color(0, DARK_GREEN_G, 0));
         }
-
-        handleOrderBookEmptyState(responseModel);
-
-        final boolean enableBuySell = currentlySelectedMarket != null
-            && currentlySelectedMarket.isBuySellEnabled()
-            && !responseModel.isReconnecting();
-
-        updateButtonStates(enableBuySell);
-    }
-
-    private void handleOrderBookEmptyState(final OrderBookResponseModel responseModel) {
-        if (responseModel.isEmpty()) {
-            if (responseModel.getMessage() != null) {
-                orderBookEmptyLabel.setText(responseModel.getMessage());
-            }
-            else {
-                orderBookEmptyLabel.setText("No orders yet");
-            }
-        }
-        else if (!responseModel.isReconnecting()) {
-            orderBookEmptyLabel.setText(EMPTY_TEXT);
-        }
     }
 
     private void updateButtonStates(final boolean enabled) {
         buyButton.setEnabled(enabled);
         sellButton.setEnabled(enabled);
         settleButton.setEnabled(enabled);
-
-        if (!enabled && currentlySelectedMarket != null && !currentlySelectedMarket.isBuySellEnabled()) {
-            statusLabel.setText("Market Closed");
-            statusLabel.setForeground(Color.GRAY);
-        }
+        placeOrderButton.setEnabled(enabled);
     }
 
     @Override
@@ -488,48 +521,15 @@ public class MarketsFrame extends JFrame implements MarketsView, SettleMarketVie
         JOptionPane.showMessageDialog(this, message, ERROR_TITLE, JOptionPane.ERROR_MESSAGE);
     }
 
-    // ---- SettleMarketView (UC6) ----
-
     @Override
     public void showSettlementResult(final String message) {
         statusLabel.setText(message);
-        String alice = "";
-        String bob = "";
-        String you = "";
-
-        try {
-            alice = "Alice: $" + StakeMateApp.getAccountRepo()
-                .findByUsername("alice").getBalance();
-            bob = "Bob:   $" + StakeMateApp.getAccountRepo()
-                .findByUsername("bob").getBalance();
-
-            final String userToFind;
-            if (currentUser != null) {
-                userToFind = currentUser;
-            }
-            else {
-                userToFind = "user";
-            }
-
-            you = "You:   $" + StakeMateApp.getAccountRepo()
-                .findByUsername(userToFind).getBalance();
-        }
-        // -@cs[IllegalCatch] Demo code needs to robustly handle missing data during presentation
-        catch (final Exception ignored) {
-            // Suppressed: It is acceptable for demo balances to fail if repo is not ready
-        }
-
-        JOptionPane.showMessageDialog(
-            this,
-            message + "\n\nAccount Balances:\n" + alice + "\n" + bob + "\n" + you,
-            "Settlement Complete",
-            JOptionPane.INFORMATION_MESSAGE);
+        JOptionPane.showMessageDialog(this, message, "Settlement Complete", JOptionPane.INFORMATION_MESSAGE);
     }
 
     @Override
     public void showSettlementError(final String errorMessage) {
-        JOptionPane.showMessageDialog(this, errorMessage,
-            "Settlement Error", JOptionPane.ERROR_MESSAGE);
+        JOptionPane.showMessageDialog(this, errorMessage, "Settlement Error", JOptionPane.ERROR_MESSAGE);
     }
 
     public CommentsPanel getCommentsPanel() {
@@ -544,138 +544,32 @@ public class MarketsFrame extends JFrame implements MarketsView, SettleMarketVie
         return currentlySelectedMarket;
     }
 
-    // ---- Order Book Table Model ----
+    /**
+     * Enables the order book popup.
+     * Kept for API compatibility with StakeMateApp.
+     */
+    public void enableOrderBookPopup() {
+        // No-op
+    }
 
-    private static final class OrderBookTableModel extends AbstractTableModel {
-
-        private static final String[] COLUMNS = {
-            "Bid Size", "Bid Price", "Ask Price", "Ask Size",
-        };
-
-        private final List<Row> rows = new ArrayList<>();
-
-        public void setOrderBook(final OrderBook orderBook) {
-            rows.clear();
-
-            final List<OrderBookEntry> bids = new ArrayList<>(orderBook.getBids());
-            final List<OrderBookEntry> asks = new ArrayList<>(orderBook.getAsks());
-
-            bids.sort(Comparator.comparingDouble(OrderBookEntry::getPrice).reversed());
-            asks.sort(Comparator.comparingDouble(OrderBookEntry::getPrice));
-
-            final int max = Math.max(bids.size(), asks.size());
-            for (int i = 0; i < max; i++) {
-                OrderBookEntry bid = null;
-                if (i < bids.size()) {
-                    bid = bids.get(i);
-                }
-
-                OrderBookEntry ask = null;
-                if (i < asks.size()) {
-                    ask = asks.get(i);
-                }
-
-                rows.add(new Row(bid, ask));
-            }
-
-            fireTableDataChanged();
+    private void openOrderBookPopup() {
+        if (currentlySelectedMarket == null) {
+            JOptionPane.showMessageDialog(this, "Select a market first.");
         }
-
-        public void clear() {
-            rows.clear();
-            fireTableDataChanged();
+        else if (currentUser == null) {
+            JOptionPane.showMessageDialog(this, "No logged-in user.");
         }
-
-        @Override
-        public int getRowCount() {
-            return rows.size();
-        }
-
-        @Override
-        public int getColumnCount() {
-            return COLUMNS.length;
-        }
-
-        @Override
-        public String getColumnName(final int column) {
-            return COLUMNS[column];
-        }
-
-        @Override
-        public Object getValueAt(final int rowIndex, final int columnIndex) {
-            final Row row = rows.get(rowIndex);
-            Object result = "";
-            switch (columnIndex) {
-                case COL_BID_QTY:
-                    result = formatVal(row.getBidQty());
-                    break;
-                case COL_BID_PRICE:
-                    result = formatVal(row.getBidPrice());
-                    break;
-                case COL_ASK_PRICE:
-                    result = formatVal(row.getAskPrice());
-                    break;
-                case COL_ASK_QTY:
-                    result = formatVal(row.getAskQty());
-                    break;
-                default:
-                    result = "";
-                    break;
-            }
-            return result;
-        }
-
-        private String formatVal(final Double val) {
-            final String result;
-            if (val == null) {
-                result = "";
+        else {
+            final String userId = StakeMateApp.userRepo.getUserIdByUsername(currentUser);
+            if (userId == null) {
+                JOptionPane.showMessageDialog(this, "Cannot find profile for: " + currentUser);
             }
             else {
-                result = "$" + String.format("%.2f", val);
-            }
-            return result;
-        }
-
-        private static final class Row {
-            private final Double bidQty;
-            private final Double bidPrice;
-            private final Double askPrice;
-            private final Double askQty;
-
-            Row(final OrderBookEntry bid, final OrderBookEntry ask) {
-                if (bid != null) {
-                    this.bidQty = bid.getQuantity();
-                    this.bidPrice = bid.getPrice();
-                }
-                else {
-                    this.bidQty = null;
-                    this.bidPrice = null;
-                }
-
-                if (ask != null) {
-                    this.askPrice = ask.getPrice();
-                    this.askQty = ask.getQuantity();
-                }
-                else {
-                    this.askPrice = null;
-                    this.askQty = null;
-                }
-            }
-
-            public Double getBidQty() {
-                return bidQty;
-            }
-
-            public Double getBidPrice() {
-                return bidPrice;
-            }
-
-            public Double getAskPrice() {
-                return askPrice;
-            }
-
-            public Double getAskQty() {
-                return askQty;
+                final var uc = StakeMateApp.getPlaceOrderUseCase();
+                final var f = new stakemate.interface_adapter.viewOrderBook.OrderBookTradingFrame(
+                    uc, userId, currentlySelectedMarket.getId());
+                f.setDefaultCloseOperation(JFrame.DISPOSE_ON_CLOSE);
+                f.setVisible(true);
             }
         }
     }
