@@ -21,6 +21,9 @@ import javax.swing.table.DefaultTableModel;
 import stakemate.interface_adapter.view_profile.ProfileState;
 import stakemate.interface_adapter.view_profile.ProfileViewModel;
 
+/**
+ * Frame for the Profile View.
+ */
 public class ProfileFrame extends JFrame implements PropertyChangeListener {
 
     private final JButton backButton = new JButton("Back");
@@ -29,12 +32,12 @@ public class ProfileFrame extends JFrame implements PropertyChangeListener {
     private final JLabel balanceLabel = new JLabel("Balance: -");
     private final JLabel pnlLabel = new JLabel("PnL: 0");
     private final String[] openColumns = {
-        "Market Name", "Team", "Buy Price", "Size", "Buy Amt ($)", "Profit if Won ($)"
+            "Market Name", "Team", "Buy Price", "Size", "Buy Amt", "Profit if Won"
     };
     private final DefaultTableModel openModel = new DefaultTableModel(openColumns, 0);
     private final JTable openTable = new JTable(openModel);
     private final String[] historyColumns = {
-        "Market Name", "Team", "Buy Price", "Size", "Profit ($)"
+            "Market Name", "Team", "Buy Price", "Size", "Profit"
     };
     private final DefaultTableModel historyModel = new DefaultTableModel(historyColumns, 0);
     private final JTable historyTable = new JTable(historyModel);
@@ -44,13 +47,34 @@ public class ProfileFrame extends JFrame implements PropertyChangeListener {
     private final JButton sortHistorySizeBtn = new JButton("Sort by Size");
 
     private ProfileViewModel viewModel;
+    private stakemate.interface_adapter.view_profile.ViewProfileController controller;
 
+    private stakemate.use_case.view_profile.SortCriteria currentOpenSort = stakemate.use_case.view_profile.SortCriteria.DATE;
+    private stakemate.use_case.view_profile.SortCriteria currentHistoricalSort = stakemate.use_case.view_profile.SortCriteria.DATE;
+
+    /**
+     * Sets the view profile controller.
+     * 
+     * @param controller the controller.
+     */
+    public void setController(final stakemate.interface_adapter.view_profile.ViewProfileController controller) {
+        this.controller = controller;
+    }
+
+    /**
+     * Constructs a ProfileFrame.
+     */
     public ProfileFrame() {
         super("StakeMate - My Profile");
         initUi();
         hookEvents();
     }
 
+    /**
+     * Sets the profile view model.
+     * 
+     * @param viewModel the view model.
+     */
     public void setViewModel(final ProfileViewModel viewModel) {
         this.viewModel = viewModel;
         this.viewModel.addPropertyChangeListener(this);
@@ -87,6 +111,14 @@ public class ProfileFrame extends JFrame implements PropertyChangeListener {
         openSortPanel.add(sortOpenDateBtn);
         openSortPanel.add(sortOpenSizeBtn);
 
+        // Center align columns for Open Table
+        final javax.swing.table.DefaultTableCellRenderer centerRenderer = new javax.swing.table.DefaultTableCellRenderer();
+        centerRenderer.setHorizontalAlignment(JLabel.CENTER);
+        openTable.setDefaultRenderer(Object.class, centerRenderer);
+        // Also center headers
+        ((javax.swing.table.DefaultTableCellRenderer) openTable.getTableHeader().getDefaultRenderer())
+                .setHorizontalAlignment(JLabel.CENTER);
+
         openPanel.add(openSortPanel, BorderLayout.NORTH);
         openPanel.add(new JScrollPane(openTable), BorderLayout.CENTER);
         final JPanel historyPanel = new JPanel(new BorderLayout());
@@ -96,6 +128,12 @@ public class ProfileFrame extends JFrame implements PropertyChangeListener {
         historySortPanel.add(new JLabel("Sort by: "));
         historySortPanel.add(sortHistoryDateBtn);
         historySortPanel.add(sortHistorySizeBtn);
+
+        // Center align columns for History Table
+        historyTable.setDefaultRenderer(Object.class, centerRenderer);
+        // Also center headers
+        ((javax.swing.table.DefaultTableCellRenderer) historyTable.getTableHeader().getDefaultRenderer())
+                .setHorizontalAlignment(JLabel.CENTER);
 
         historyPanel.add(historySortPanel, BorderLayout.NORTH);
         historyPanel.add(new JScrollPane(historyTable), BorderLayout.CENTER);
@@ -114,26 +152,81 @@ public class ProfileFrame extends JFrame implements PropertyChangeListener {
     private void hookEvents() {
         backButton.addActionListener(e -> setVisible(false));
         sortOpenDateBtn
-            .addActionListener(e -> JOptionPane.showMessageDialog(this, "Sort by Date not implemented yet."));
+                .addActionListener(e -> {
+                    final String username = viewModel.getState().getUsername();
+                    if (username != null) {
+                        currentOpenSort = stakemate.use_case.view_profile.SortCriteria.DATE;
+                        controller.execute(username, currentOpenSort, currentHistoricalSort);
+                    }
+                });
         sortOpenSizeBtn
-            .addActionListener(e -> JOptionPane.showMessageDialog(this, "Sort by Size not implemented yet."));
+                .addActionListener(e -> {
+                    final String username = viewModel.getState().getUsername();
+                    if (username != null) {
+                        currentOpenSort = stakemate.use_case.view_profile.SortCriteria.SIZE;
+                        controller.execute(username, currentOpenSort, currentHistoricalSort);
+                    }
+                });
         sortHistoryDateBtn
-            .addActionListener(e -> JOptionPane.showMessageDialog(this, "Sort by Date not implemented yet."));
+                .addActionListener(e -> {
+                    final String username = viewModel.getState().getUsername();
+                    if (username != null) {
+                        currentHistoricalSort = stakemate.use_case.view_profile.SortCriteria.DATE;
+                        controller.execute(username, currentOpenSort, currentHistoricalSort);
+                    }
+                });
         sortHistorySizeBtn
-            .addActionListener(e -> JOptionPane.showMessageDialog(this, "Sort by Size not implemented yet."));
+                .addActionListener(e -> {
+                    final String username = viewModel.getState().getUsername();
+                    if (username != null) {
+                        currentHistoricalSort = stakemate.use_case.view_profile.SortCriteria.SIZE;
+                        controller.execute(username, currentOpenSort, currentHistoricalSort);
+                    }
+                });
+    }
+
+    private javax.swing.Timer autoRefreshTimer;
+
+    private void initAutoRefresh() {
+        // Refresh every 3 seconds
+        autoRefreshTimer = new javax.swing.Timer(3000, e -> {
+            if (isVisible() && viewModel != null && controller != null) {
+                final String username = viewModel.getState().getUsername();
+                if (username != null && !username.isEmpty()) {
+                    controller.execute(username, currentOpenSort, currentHistoricalSort);
+                }
+            }
+        });
+    }
+
+    @Override
+    public void setVisible(final boolean b) {
+        super.setVisible(b);
+        if (b) {
+            if (autoRefreshTimer == null) {
+                initAutoRefresh();
+            }
+            autoRefreshTimer.start();
+        } else {
+            if (autoRefreshTimer != null) {
+                autoRefreshTimer.stop();
+            }
+        }
     }
 
     @Override
     public void propertyChange(final PropertyChangeEvent evt) {
         final ProfileState state = (ProfileState) evt.getNewValue();
         if (state.getError() != null) {
-            JOptionPane.showMessageDialog(this, state.getError(), "Error", JOptionPane.ERROR_MESSAGE);
+            // Don't show error dialog on auto-refresh to avoid spamming user
+            // JOptionPane.showMessageDialog(this, state.getError(), "Error",
+            // JOptionPane.ERROR_MESSAGE);
             return;
         }
 
         usernameLabel.setText("Username: " + state.getUsername());
-        balanceLabel.setText("Balance: " + state.getBalance());
-        pnlLabel.setText("PnL: " + state.getPnl());
+        balanceLabel.setText(String.format("Balance: $%.2f", state.getBalance()));
+        pnlLabel.setText(String.format("PnL: $%.2f", state.getPnl()));
 
         updateTable(openModel, state.getOpenPositions());
         updateTable(historyModel, state.getHistoricalPositions());
